@@ -10,8 +10,10 @@
 		}
 	],
 	"require": {
-		"php": ">=5.3.2",
-		"nette/nette": "~2.1@dev"
+		"php": ">=5.4",
+		"nette/utils": "~2.2",
+		"nette/forms": "~2.2",
+		"nette/application": "~2.2"
 	},
 	"autoload": {
 		"psr-4": {
